save@2023-09-03 19:31+9:00

+++ qujila_prisma/src/model/float.rs
@@ -1,11 +1,11 @@
 pub struct FloatField {
-    pub default: Option<i64>,
+    pub default: Option<f64>,
 }
 
 pub struct FloatListField {
-    pub default: Option<Vec<i64>>,
+    pub default: Option<Vec<f64>>,
 }
 
 pub struct FloatOptionalField {
-    pub default: Option<i64>,
+    pub default: Option<f64>,
 }
